requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
loguru==0.7.0
magic-filter==1.0.12
multidict==6.2.0
openpyxl==3.1.2
peewee==3.16.2
pillow==10.4.0
propcache==0.3.1
python-dotenv==1.0.0
pytz==2025.2
yarl==1.18.3
